requirements.txt
statistics >= 1.0.3.5
git+https://github.com/Rapptz/discord.py
python-dotenv >= 0.20.0
